+++ diesel/src/expression/functions/date_and_time.rs
@@ -1,5 +1,5 @@
 use backend::Backend;
-use expression::{Expression, NonAggregate};
+use expression::Expression;
 use query_builder::*;
 use result::QueryResult;
 use sql_types::*;
@@ -7,14 +7,12 @@ use sql_types::*;
 /// Represents the SQL `CURRENT_TIMESTAMP` constant. This is equivalent to the
 /// `NOW()` function on backends that support it.
 #[allow(non_camel_case_types)]
-#[derive(Debug, Copy, Clone, QueryId)]
+#[derive(Debug, Copy, Clone, QueryId, AppearsOnTable, SelectableExpression, NonAggregate)]
 pub struct now;
 
 impl Expression for now {
     type SqlType = Timestamp;
 }
-
-impl NonAggregate for now {}
 
 impl<DB: Backend> QueryFragment<DB> for now {
     fn walk_ast(&self, mut out: AstPass<DB>) -> QueryResult<()> {
@@ -22,8 +20,6 @@ impl<DB: Backend> QueryFragment<DB> for now {
         Ok(())
     }
 }
-
-impl_selectable_expression!(now);
 
 operator_allowed!(now, Add, add);
 operator_allowed!(now, Sub, sub);