+++ diesel/src/expression/functions/mod.rs
@@ -14,7 +14,7 @@ macro_rules! sql_function_body {
         $helper_ty_docs:expr
     ) => {
         #[allow(non_camel_case_types)]
-        #[derive(Debug, Clone, Copy, QueryId)]
+        #[derive(Debug, Clone, Copy, QueryId, SelectableExpression, AppearsOnTable, NonAggregate)]
         #[doc(hidden)]
         pub struct $struct_name<$($arg_name),*> {
             $($arg_name: $arg_name),*
@@ -56,27 +56,6 @@ macro_rules! sql_function_body {
                 out.push_sql(")");
                 Ok(())
             }
-        }
-
-        #[allow(non_camel_case_types)]
-        impl<$($arg_name),*, QS> $crate::expression::SelectableExpression<QS> for $struct_name<$($arg_name),*> where
-            $($arg_name: $crate::expression::SelectableExpression<QS>,)*
-            $struct_name<$($arg_name),*>: $crate::expression::AppearsOnTable<QS>,
-        {
-        }
-
-        #[allow(non_camel_case_types)]
-        impl<$($arg_name),*, QS> $crate::expression::AppearsOnTable<QS> for $struct_name<$($arg_name),*> where
-            $($arg_name: $crate::expression::AppearsOnTable<QS>,)*
-            $struct_name<$($arg_name),*>: $crate::expression::Expression,
-        {
-        }
-
-        #[allow(non_camel_case_types)]
-        impl<$($arg_name),*> $crate::expression::NonAggregate for $struct_name<$($arg_name),*> where
-            $($arg_name: $crate::expression::NonAggregate,)*
-            $struct_name<$($arg_name),*>: $crate::expression::Expression,
-        {
         }
     }
 }
@@ -600,18 +579,12 @@ macro_rules! no_arg_sql_function_body_except_to_sql {
     ($type_name:ident, $return_type:ty, $docs:expr) => {
         #[allow(non_camel_case_types)]
         #[doc=$docs]
-        #[derive(Debug, Clone, Copy, QueryId)]
+        #[derive(Debug, Clone, Copy, QueryId, SelectableExpression, AppearsOnTable, NonAggregate)]
         pub struct $type_name;
 
         impl $crate::expression::Expression for $type_name {
             type SqlType = $return_type;
         }
-
-        impl<QS> $crate::expression::SelectableExpression<QS> for $type_name {}
-
-        impl<QS> $crate::expression::AppearsOnTable<QS> for $type_name {}
-
-        impl $crate::expression::NonAggregate for $type_name {}
     };
 }