+++ diesel/src/expression/grouped.rs
@@ -1,9 +1,9 @@
 use backend::Backend;
-use expression::{Expression, NonAggregate};
+use expression::Expression;
 use query_builder::*;
 use result::QueryResult;
 
-#[derive(Debug, Copy, Clone, QueryId, Default, DieselNumericOps)]
+#[derive(Debug, Copy, Clone, QueryId, Default, DieselNumericOps, AppearsOnTable, SelectableExpression, NonAggregate)]
 pub struct Grouped<T>(pub T);
 
 impl<T: Expression> Expression for Grouped<T> {
@@ -17,12 +17,4 @@ impl<T: QueryFragment<DB>, DB: Backend> QueryFragment<DB> for Grouped<T> {
         out.push_sql(")");
         Ok(())
     }
-}
-
-impl_selectable_expression!(Grouped<T>);
-
-impl<T: NonAggregate> NonAggregate for Grouped<T>
-where
-    Grouped<T>: Expression,
-{
 }