+++ diesel/src/pg/expression/array_comparison.rs
@@ -1,5 +1,5 @@
 use expression::subselect::Subselect;
-use expression::{AsExpression, Expression, NonAggregate};
+use expression::{AsExpression, Expression};
 use pg::Pg;
 use query_builder::*;
 use result::QueryResult;
@@ -64,7 +64,7 @@ where
 }
 
 #[doc(hidden)]
-#[derive(Debug, Copy, Clone, QueryId)]
+#[derive(Debug, Copy, Clone, QueryId, AppearsOnTable, SelectableExpression, NonAggregate)]
 pub struct Any<Expr> {
     expr: Expr,
 }
@@ -94,16 +94,8 @@ where
     }
 }
 
-impl_selectable_expression!(Any<Expr>);
-
-impl<Expr> NonAggregate for Any<Expr>
-where
-    Expr: NonAggregate,
-{
-}
-
 #[doc(hidden)]
-#[derive(Debug, Copy, Clone, QueryId)]
+#[derive(Debug, Copy, Clone, QueryId, AppearsOnTable, SelectableExpression, NonAggregate)]
 pub struct All<Expr> {
     expr: Expr,
 }
@@ -131,14 +123,6 @@ where
         out.push_sql(")");
         Ok(())
     }
-}
-
-impl_selectable_expression!(All<Expr>);
-
-impl<Expr> NonAggregate for All<Expr>
-where
-    Expr: NonAggregate,
-{
 }
 
 pub trait AsArrayExpression<ST> {